backend/requirements.txt
flask
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
flask-cors